+++ src/components/App/index.tsx
@@ -4,7 +4,7 @@ import Hand from '@components/Hand';
 import HandIcon from '@components/HandIcon';
 import NotificationQueue from '@components/NotificationQueue';
 import PlayArea from '@components/PlayArea';
-import { Actions } from '@utilities/types';
+import { Actions, Icons } from '@utilities/types';
 import { Context, getDefaultState, Reducer } from './store';
 import Style from './style';
 
@@ -13,18 +13,23 @@ const App = (): JSX.Element => {
   const [state, dispatch] = React.useReducer(Reducer, getDefaultState());
   const { shouldShowHand } = state;
 
-  const handleClick = () => {
+  const handleAppClick = () => {
     if (activeCard > -1) setActiveCard(-1);
     if (shouldShowHand) dispatch({ type: Actions.HIDE_HAND });
   };
 
+  const handleCardClick = (id: number) => {
+    if (id === activeCard) id = -1;
+    setActiveCard(id);
+  };
+
   return (
-    <Style className='app' onClick={handleClick}>
+    <Style className='app' onClick={handleAppClick}>
       <Context.Provider value={{ dispatch, state }}>
-        <PlayArea activeCard={activeCard} setActiveCard={setActiveCard} />
-        <ArrowIcon iconType='END_TURN' />
+        <PlayArea activeCard={activeCard} handleCardClick={handleCardClick} />
+        <ArrowIcon iconType={Icons.END_TURN} />
         <HandIcon />
-        <Hand />
+        <Hand activeCard={activeCard} handleCardClick={handleCardClick} />
         <NotificationQueue />
       </Context.Provider>
     </Style>