+++ docs/changelog.md
@@ -10,6 +10,8 @@ Ends October 14, 2020
 - update cron instructions in cron.md.
 - conditionally set notification type as either red or green.
 - add charset to index.html.
+- add noImplicitAny to tsconfig.
+- give Hand Cards the abiliity to be active.
 
 ## Sprint 4
 Ends September 16, 2020